feat!: Adding the audio functionality in the meditation section of the app
Initializing the sound when clicked on the "Start Meditation" CustomButton, on clicking, it setMeditating to true and calls the toggleMeditationSessionStatus function which setSeconds to 10 if remaining seconds are zero, and setMeditating to opposite of it's current state and then toggle the sound.

But there's a possible bug that when clicking on the button first time, the music starts playing and as soon as we click on the Pause button, it stops but when we again click on the Resume button it doesn't start playing, idk why ;).

Also, refactored the way CustomButton title and onPress functionalities works.