feat: Fix bug preventing audio playback in Meditate/[id] screen
Initialize the audio sound when the "Start Meditation" button is clicked. Set the meditating state to true and call the toggleMeditationSessionStatus function. If the remaining seconds are zero, set the seconds to 10. Also, refactor the CustomButton title and onPress functionalities.

Addresses a bug where the audio playback doesn't start after clicking the Resume button for the second time.

Refactor the Meditate/[id] screen to improve the appearance and behavior of the start button. Allow the user to reset the timer after it reaches zero. Apply dynamic styles to the button based on the meditation state.