chore: Update app.json and eas.json
This commit updates the app.json file to add the bundleIdentifier and package fields for iOS and Android respectively. It also updates the eas.json file to configure the CLI version and build and submit options. These changes enhance the app's configuration and compatibility with the latest tools and libraries.